Remove old entity-component system

+++ Library/source/EntityFramework/Types.h
@@ -3,6 +3,6 @@
 
 using ptr_t = void*;
 using cptr_t = void const*;
-using EntityID = uint32_t;
+using EntityID = uint64_t;
 using ComponentTypeID = uint16_t;
 using ByteStream = std::vector<uint8_t>;
